+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- BulkInsert: Catch `RuntimeException` from underlying *phlib/db* for deadlocks.
+  Migration from *phlib/db* mistakenly created a new `RuntimeException` class.
 
 ## [1.0.1] - 2017-04-19
 ### Fixed